Create background.rst
2.2 Data Preparation
===================================================================================================================================================
Monitoring and assessing the wetland and basin conditions, extents and trends require data preparation and computation integrity. 
This depends on the use of desired data sets, prepared and visualized in a format that is easily interpreted. The outputs generated depends on the input data, thus, it is vital to input a well-preprocessed data in order to visualize a favorable output.

This report is therefore intended to provide a guide to the data preparation and outline the processing chains requirements and visualization procedures. 
Wetland monitoring and assessment is dependent on the satellite imagery products termed as sub indicators that are computed from the combination of specific multi-spectral bands. The desired band combination is used in computations to produce the desired output. The desired products are then uploaded into the WeMAST GeoServer, whereby they are made accessible to the client upon request, in form of Geovisualised maps, charts, time series data, statistics as well as tabulated Metadata. 

Few in-situ computations are however applied in the backend system to generate real time monitoring and assessment results such as burnt areas, flooded zoned areas, vegetation cover and land use land cover change maps based on the availability and temporal frequency of the datasets.
The WeMAST geoportal system is envisioned to provide a monitoring and assessment baseline information for the 4 major basins together with the existing wetland areas within the sub basins. The basins are the Zambezi, Cuvelai, Limpopo and Okavango.

The sub - indicators applied in monitoring wetland and basin conditions are largely categorized into three, in order to capture the sustainable use of the basin resources while assessing their sensitivity, exposure as well as resilience on their continuous use.
The sub indicators that are used to study the above conditions are listed below, with their procedures discussed in the 
`Data Preparation Manual <https://drive.google.com/file/d/1yDdZSvnr14I5uU0nMDh77FSQfpqN5N50/view?usp=sharing/>`_.